middleware/job/common.py
from flask_restful import Resource, abort


class JobApi(Resource):
    '''API for direct manipulation of jobs'''
    def __init__(self, **kwargs):
        # Inject job service
        self.jobs = kwargs['job_repository']

    def abort_if_not_found(self, job_id):
        if(not self.jobs.exists(job_id)):
            abort(404, message="Job {} not found".format(job_id))

    def get(self, job_id):
        self.abort_if_not_found(job_id)
        job = self.jobs.get_by_id(job_id)
        return job, 200
